Made KanbanBoard Generic
