+++ frontend/src/pages/components/House.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import DraggableHouse from "./DraggableHouse";
+import DraggableItem from "./DraggableItem";
 import HouseDesign from "./HouseDesign";
 
-export default function House(props) {
+export default function House(props, current) {
     return (
         <Link to={`/home/:${props.id}`}>
-            <DraggableHouse id={props.id} current={props.current}>
-                <HouseDesign address={props.address} />
-            </DraggableHouse>
+            <DraggableItem id={props.id} current={current}>
+                <HouseDesign name={props.address} />
+            </DraggableItem>
         </Link>
     )
 }
